Add files via upload
JOURNEY FANCY FLYING ENABLER FOR KBM

> DISCLAIMER!!

i made this tool purely for fun, and to allow keyboard-only players (like me) to experience/experiment with some of the more advanced Fancy Flying tech which is normally inaccessible to those of us who do not own a controller or simply prefer playing with a mouse and keyboard. Since the tool allows for perfect precision and consistency, please do not use it to submit any speedrun times, as one might argue it could be used to gain an unfair advantage in certain situations and that is certainly not the reason why i made this. This tool is meant for accessibility and goofing around in the desert sky. Please use it responsibly and respectfully towards fellow companions. Thank you.

> PREREQUISITES AND USAGE

The script runs in Python 3. It also utilizes the 'vgamepad' and 'keyboard' modules which can be found here:
vgamepad: https://pypi.org/project/vgamepad/
keyboard: https://pypi.org/project/keyboard/

To activate the script, simply open the .py file or run it through your Python shell. All the script does is emulate a virtual gamepad and send left-stick inputs to your computer, so the game does not need to be running beforehand nor does the script directly interact with it in any way. Additionally, the game recognize both gamepad and keyboard inputs at the same time and the script doesn't overwrite anything, so all your normal keyboard inputs will be completely unchanged.
Once you launch it, you can check that it's working correctly here: https://hardwaretester.com/gamepad . Simply press any of the binds and look at the stick move.

> CONTROLS
By default:
ESC ends the script
'Q' moves the left stick forward 60%
'E' moves the left stick forward 85%

All keybinds and percentages can be customized to whatever you like. At the start of the script there's a small labeled section where you can change everything to suit your needs. i decided to set these two values to 60% and 85% because they allow you to perform Flap Boost and Infinite Boost, but if you find values that work better for you feel free to change them to whatever you want.

Aaand that's pretty much it! Once again please use this tool responsibly and have fun.
See you in the sky.

-Tofu